Add implicitly added methods to entity documentation
